services/platform/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.2
bcrypt==4.1.1
pyjwt==2.8.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
